Fix 2 issues, add retry system.
Fixed an issue where the program would stop working if the number of
detected embeds was divisible by 10, but also greater than 10 (20, 30,
40, etc.). This was caused by the system trying to send a null message
to Discord.

Fixed an issue where the program would stop working after a long time.
This was caused by the program recycling the old HTTP Request, instead
of creating a new one. As a result, every request resulted in an
additional "timestamp" parameter being passed, eventually leading too a
URI too long error from the server.

Added a retry system. If a request fails, the program will try to get
the response 2 more times (if necessary), then parse it into a GSON
object again.